requirements.txt
langchain>=1.0,<2.0
langchain-core>=1.0,<2.0
langgraph>=1.0,<2.0
langchain-anthropic
langsmith>=0.1.99
python-dotenv
